Don't use pipe symlink on client (#83)
The named pipe prefix `\\.\pipe` is a OS user specific symlink in the kernel object namespace which points to `\Device\NamedPipe`. In order to prevent an non-admin app, running in the same OS user context as the client browser, from redirecting this symlink elsewhere, the client will always use the pipe prefix `\Device\NamedPipe` to build the full path to the named pipe. Using the latter path means the client must call `NtCreateFile` instead of the win32 `CreateFile`.
---------
Co-authored-by: Roger Tawa <rogerta@chromium.org>