Delay per request.
Add ability to specify a response delay per request.  For text
and file requests, the body to be scanned needs to have the string
"delay=x" where x is the delay in seconds.  For print requests,
the URL should have a parameter of the form "delay=x".

To test, run the demo agent as follows:

   agent --user --queued

Run the demo client as follows (sends two requests, one whose
response is delayed 7 seconds, the other not delyaed):

   browser --user --threaded "allow delay=7" "block"

Test print use cases with chrome.

This commit includes some tiddying up of the output in the demo
agent.  It also includes a fix in the demo client that assumed
responses were sent in the same order as the requests.